email_agent/requirements.txt
smolagents
python-dotenv
requests
